refactor(onboard): split tests under architecture budget
